Refactor Talon wrapper member variables/wrap more methods.
Specifically the set method, where it is desierable to specify
ControlMode explicitly. If we don't override it then our state may
diverge from the actual state of the Talon we're wrapping. I suspect
that there are more methods on the superclass that will screw with our
state, but this is the most used/important by far.